+++ src/main/java/com/prgrms/ohouse/domain/user/model/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.prgrms.ohouse.domain.user.model;
+package com.prgrms.ohouse.domain.user.model.exception;
 
 public class UserNotFoundException extends RuntimeException {
